backend/requirements.txt
Flask
Flask-Cors
flask-limiter
requests
psycopg2-binary
python-dotenv
